Update README.md - clarify installation  and folder structure sections of Readme
A few readme changes to better guide the user toward a successful first load: 

Updated installation section to clarify the need for the IGDB key to run the application and setting up folder structure. Failing to do these things before spinning up Romm prevents scanning from working, so they are best to handle first. Additionally, the folder structure section is updated to call the user's attention to the list of platform names and that custom names may be used in conjunction with entries in the config.yml file.