added a maximum length to file extensions
found that the code errors when there is a period in a filename (such as hidden files in UNIX OSes) and the file doesn't have a proper extension. Apple is really bad about putting these kinds of files in. To fix it, I added a new variable to the _exclude_files function, max_ext_length and set it to default to 4. That way if another function needs a different max it can override the default, but it shouldn't require other code changes.