Delete frontend/assets/platforms/palm-os.ico
I have better ideas of what to do with this one. Since with previous mobile platforms, I've had an example phone with the logo on the screen, I want to do that for Palm as well. May also come back at some point and do that for J2ME if I have time but I think with that one, the java logo works as a platform icon for now. Here, I'm less certain.