fix: Use filename without tags or extension when matching unmatched game
If a game is manually unmatched, we update its `name` field to
`file_name` (which includes both tags and extension). After that, if we
go to the Rom list, and click on "Manual match" for that rom, the search
field is populated with that Rom name, which commonly won't retrieve any
results from the metadata providers.

This change makes the "Manual match" search term to use the filename
without tags or extension, when the Rom is currently not matched. It
also triggers the search automatically when a search term was set.

Potential fix for #1167.