fix: Show source image on each result for Manual match
The source image was not being shown on the results of the manual match
dialog, because scrapper image names do not match the source names. To
avoid this issue, the `MatchedSource` type was updated to include a
`logo_path` property, which will be used to set the correct image path.