fix: Speed up scan for matched roms without filesystem changes
When the scanning process fiends that a rom should not be scanned, it
was still updating the rom data in the database.

This database update is only needed when the existing data is different
to the newly found filesystem data, so we can avoid most of the updates
for already existing library roms.