feat: Include all EmulatorJS assets (cores, scripts) in docker image
The browser caching does not work for original emulatorjs CDN, also on new emulatorjs release the emulator integration in old versions of romm become outdated and not working anymore. This forces users who using the emulator in romm to switch to alpha/rc versions to get it work.

This update embeds all EmulatorJS assets in romm docker image and not depend from CDN anymore.