feat: Experimental PSP emulation
EmulatorJS version `4.2.0` supports PPSSPP core for PSP emulation.

This change enables PSP emulation in experimental mode. It requires
the browser to support `SharedArrayBuffer` for multi-threading support.

Currently, RomM is not Cross-Origin Isolated, which is required for
browsers to enable `SharedArrayBuffer`. In the meantime, Chrome can be
executed using the following command, to enable `SharedArrayBuffer`:

```shell
chrome --enable-features=SharedArrayBuffer
```