fix: Make nginx configuration files writable by everyone
As the container entrypoint runs `envsubst` to replace environment
variables in the nginx configuration files, the `/etc/nginx/conf.d` and
its contents must be writable by everyone.

This is needed because a user can set a custom UID/GID to run the
container, and the `envsubst` command will run as that user.