fix: Download for single-file games in folder
When trying to download a game that is contained in a folder, but is
single-file (what we consider "non-multi" game), the download would
fail, trying to make nginx proxy the folder instead of that single file.

This commit removes custom logic related to `multi` games, so the logic
ends up executing the `if len(files) == 1` block, which is the correct
one for single-file games.

A simple way to reproduce this issue is to create a game folder with a
single file inside it, and try to download or run EmulatorJS with that
game. Both `DEV_MODE=true` and `DEV_MODE=false` were failing in this
scenario.

Fixes #1644.