fix: Make default ROMM_AUTH_SECRET_KEY consistent between workers
When the `ROMM_AUTH_SECRET_KEY` environment variable is not set, and the
Gunicorn application server spawns multiple worker processes, each
worker was generating a different random secret key. This led to
intermittent authentication failures, hard to debug and reproduce.

To fix this, we now set a default secret key that is consistent across
all workers. For security reasons, it is still recommended to set a
unique and secure key via the `ROMM_AUTH_SECRET_KEY` environment
variable. When not set, the default key is now derived from the
Docker's `HOSTNAME` environment variable, or a hardcoded fallback value.