misc: Add PlayBtn component to unify button logic
The Play button logic was duplicated in three different places,
including when to run either EmulatorJS or Ruffle.

This change centralizes that logic in a new `PlayBtn` component, which
internally decides which player to use, and binds to any arguments
passed to it.

Now all three places also consider platform versions when checking
for emulation support.